Add support for downloads handled by browser
